+++ src/adapters/ui/SceneFureai.tsx
@@ -4,7 +4,6 @@ import { OverlayFureai } from './OverlayFureai'
 import { FureaiExitButton } from './FureaiExitButton'
 import { PromptInput } from './PromptInput'
 import { HeartEffect } from './HeartEffect'
-import { CharacterNameEditor } from './CharacterNameEditor'
 import { useEventBusCallback } from './hooks/useEventBus'
 import type { FeedingSuccessEvent } from '../three/FeedingInteractionAdapter'
 
@@ -19,7 +18,6 @@ export function SceneFureai(): JSX.Element {
   return (
     <>
       <OverlayFureai />
-      <CharacterNameEditor />
       <FureaiExitButton />
       <PromptInput />
       <HeartEffect triggerKey={heartKey} />